requirements.txt
# ─── Kishaan Deepak — Python dependencies ─────────────────────────────────────
# Install with: pip install -r requirements.txt

# Web framework
flask>=2.3
flask-cors>=4.0

# ML / numerical
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
scikit-image>=0.21
joblib>=1.3

# Computer vision
opencv-python-headless>=4.8

# HTTP / Werkzeug (pinned transitively by Flask — listed for clarity)
werkzeug>=2.3

# LLM / chatbot
groq>=0.10

# Config / environment
python-dotenv>=1.0

# Production WSGI server (optional — use instead of flask dev server)
gunicorn>=21.0
